pcre2 support
pcre is the native simh regex support (yay!). Add support for pcre2
because pcre2 is actively maintained. pcre is in long term maintainance
mode (yes, it works but there will be a request for this someday.)

Other changes:

sim_encode_quoted_string: Change size parameter from uint32 to size_t
    to match pcre2 lengths. Should not impact pcre support.

fprint_buffer_string: Change size parameter from uint32 to size_t.
    Impacts the SEND structure's insoff and extoff members.

struct EXPECT:
    size:     int32  -> size_t
    buf_ins:  uint32 -> size_t
    buf_data: uint32 -> size_t

struct SEND:
    insoff:   int32  -> size_t
    extoff:   int32  -> size_t

sim_defs.h: Define portable printf format prefixes for size_t, t_value
    and t_svalue. Quash a printf format warning in I7000/i7080_sys.c as
    a result.

+++ scp.h
@@ -254,8 +254,8 @@ CONST char *get_range (DEVICE *dptr, CONST char *cptr, t_addr *lo, t_addr *hi,
     uint32 rdx, t_addr max, char term);
 t_stat sim_set_environment (int32 flag, CONST char *cptr);
 t_stat sim_decode_quoted_string (const char *iptr, uint8 *optr, uint32 *osize);
-char *sim_encode_quoted_string (const uint8 *iptr, uint32 size);
-void fprint_buffer_string (FILE *st, const uint8 *buf, uint32 size);
+char *sim_encode_quoted_string (const uint8 *iptr, size_t size);
+void fprint_buffer_string (FILE *st, const uint8 *buf, size_t size);
 t_value strtotv (CONST char *cptr, CONST char **endptr, uint32 radix);
 t_svalue strtotsv (CONST char *inptr, CONST char **endptr, uint32 radix);
 int Fprintf (FILE *f, const char *fmt, ...) GCC_FMT_ATTR(2, 3);
